Move trace class fetching logic to command module
